requirements.txt
fastapi[all]
discord
flask
numpy
Pyrebase4
pydantic
requests
py-cord
firebase-admin
python-dotenv
setuptools
google
google-generativeai